♻️ : UsingTopItem 배경색 util함수로 분리
